Added logging
All of my code is commented, but the general premise for logging is as follows:
Anytime you want to log an action, call the 'log' helper. This takes the time, the action the user took (i.e the path), the username and whether he has edit rights or not and stores it into 'log.txt' as a HTML table record.
Anytime you want to access the logs, go to '/log'. This then outputs the contents of 'log.txt' directly into a HTML table.

For some reason, my environment at home is completely FUBAR so I have to check that this *actually* works when I get into uni. Syntactically, everything should be fine since it goes through a ruby interpreter without a hitch. But idk if the dbs stuff works for certain, so don't submit this version!

Also, I think 'authorised' and 'protected' helpers check the "admincontrols" thing so that responsibility of mine is already covered by the practical